Rename default.py into config.py
